soft delete part implemented
Delete part was implemented.

Note: The logic to list parts and create parts was changed, if a part is deleted then we keep the location id for record. This means that when we try to create a aporte we need to also check wich parts have deleted at == null so that we can proprely check if the location is full or no. In the future, we can rever this to simples when soft delete a part, then de location id and etc gets changed and when creating/listing parts we dont have to care about soft deleted parts